requirements.txt
Django==4.1.6
django-cors-headers==3.13.0
django-storages==1.13.2
Pillow==9.4.0
google-auth
google-cloud-storage
djangorestframework==3.14.0
